Fixing compile warnings

+++ src/glib/concurrent/posix/thread_posix.cpp
@@ -1,7 +1,7 @@
 /**
  * Copyright (c) 2015, Jozef Stefan Institute, Quintelligence d.o.o. and contributors
  * All rights reserved.
- * 
+ *
  * This source code is licensed under the FreeBSD license found in the
  * LICENSE file in the root directory of this source tree.
  */
@@ -160,7 +160,7 @@ void * TThread::EntryPoint(void * pArg) {
     try {
         pThis->Run();
     } catch (const PExcept& Except) {
-        printf("Exception in thread %ld: %s\n", pThis->GetThreadId(), Except->GetMsgStr().CStr());
+        printf("Exception in thread %ld: %s\n", (long)pThis->GetThreadId(), Except->GetMsgStr().CStr());
     } catch (...) {
         printf("Unknown exception while running thread: %s!\n", TUInt64::GetStr(pThis->GetThreadId()).CStr());
     }
